requirements.txt
einops==0.7.0
matplotlib==3.6.2
networkx==3.1
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
pandas==1.3.5
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.11.4
sympy==1.12
torch==2.0.1
torchquad==0.4.0
torchvision==0.15.2
torcheval==0.0.7
tqdm==4.64.1
transformers==4.25.0
